container views: do not report dirty children when container itself is invisible (alpha==0)
This is important especially when there are multiple timing-incompatible scrollers with timing priority defined,
but only one is visible. The invisible ones must definitely not report dirty children, as this would
trigger display updates for changes that are not visible, but will mess up timing of the visible
scroller.
Of course, this applies to any container view.